Safari-611.1.10/LayoutTests/http/tests/security/sandboxed-iframe-geolocation-watchPosition-expected.txt

Tests that navigator.geolocation.watchPosition() returns error PERMISSION_DENIED when called from a document in a sandboxed iframe.



--------
Frame: '<!--frame1-->'
--------
PASS error.code is error.PERMISSION_DENIED.
PASS error.message is "Origin does not have permission to use Geolocation service".
